Set variable Decl for identifier
Issue:
https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/IC1QG4

Description:
UI feedback: When the type of property a_ in the class is a lambda, and this.a() is used to call this lambda, the Decl is not set for the variable without an identifier.
On the UI side, the corresponding declaration cannot be obtained through getDeclarationFromIdentifier.

Tested-by:
ninja tests (passed)
ets_testrunner (passed)

Signed-off-by: zengzengran <zengran2@huawei.com>#